Revert Tailwind Integration
