+++ web/.stylelintrc.json
@@ -1,11 +1,10 @@
 {
-  "extends": ["stylelint-config-tailwindcss/scss", "stylelint-config-standard-scss"],
+  "extends": "stylelint-config-standard-scss",
   "rules": {
     "selector-class-pattern": null,
     "custom-property-pattern": null,
     "no-descending-specificity": null,
     "scss/no-global-function-names": null,
-    "scss/function-no-unknown": null,
     "selector-pseudo-class-no-unknown": [
       true,
       {
