--- web/libs/storybook/.storybook/preview.scss
@@ -1,3 +1,0 @@
-@import '../../ui/src/tokens/colors';
-@import '../../ui/src/tokens/typography';
-
