--- web/libs/storybook/.storybook/preview.ts
@@ -1,5 +1,0 @@
-import "@humansignal/ui/tailwind.css";
-import "./preview.scss";
-
-export const parameters = {};
-export const tags = ["autodocs", "autodocs"];
