--- web/libs/storybook/select/select.stories.tsx
@@ -1,42 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import {
-  Select,
-  SelectTrigger,
-  SelectValue,
-  SelectContent,
-  SelectLabel,
-  SelectItem,
-  SelectGroup,
-} from "../../ui/src/shad/components/ui/select.tsx";
-
-const meta: Meta<typeof Select> = {
-  component: Select,
-  render: ({ form, ...args }) => {
-    return (
-      <Select>
-        <SelectTrigger className="w-[180px]">
-          <SelectValue placeholder="Select a fruit" />
-        </SelectTrigger>
-        <SelectContent>
-          <SelectGroup>
-            <SelectLabel>Fruits</SelectLabel>
-            <SelectItem value="apple">Apple</SelectItem>
-            <SelectItem value="banana">Banana</SelectItem>
-            <SelectItem value="blueberry">Blueberry</SelectItem>
-            <SelectItem value="grapes">Grapes</SelectItem>
-            <SelectItem value="pineapple">Pineapple</SelectItem>
-          </SelectGroup>
-        </SelectContent>
-      </Select>
-    );
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof Select>;
-
-export const Primary: Story = {
-  args: {
-    value: "Apple",
-  },
-};
