+++ web/libs/ui/.storybook/preview.scss
@@ -1,0 +1,2 @@
+@import '../src/tokens/colors.scss';
+@import '../src/tokens/typography.scss';
