+++ web/libs/ui/.storybook/preview.ts
@@ -1,0 +1,3 @@
+import "./preview.scss";
+
+export const parameters = {};
