--- web/libs/ui/components.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "https://ui.shadcn.com/schema.json",
-  "style": "default",
-  "rsc": false,
-  "tsx": true,
-  "tailwind": {
-    "config": "../../tailwind.config.ts",
-    "css": "./src/tailwind.css",
-    "baseColor": "slate",
-    "cssVariables": true,
-    "prefix": ""
-  },
-  "aliases": {
-    "components": "@humansignal/shad/components",
-    "utils": "@humansignal/shad/utils"
-  }
-}
