+++ web/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "ui:serve": "nx storybook storybook",
+    "ui:serve": "nx storybook ui",
     "ui:test:component": "nx test-component ui",
     "lint": "biome check --apply .",
     "lint-scss": "yarn stylelint '**/*.scss' --fix",
@@ -42,13 +42,10 @@
   "private": true,
   "dependencies": {
     "@martel/audio-file-decoder": "^2.3.15",
-    "@radix-ui/react-select": "^2.1.4",
-    "@radix-ui/react-slot": "^1.1.1",
     "@radix-ui/react-toast": "^1.1.5",
     "@thi.ng/rle-pack": "^3.1.30",
     "antd": "^4.3.3",
     "chroma-js": "^2.1.1",
-    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "codemirror": "^5.59.4",
     "d3": "^5.16.0",
@@ -65,7 +62,6 @@
     "lodash.get": "^4.4.0",
     "lodash.ismatch": "^4.4.0",
     "lodash.throttle": "^4.1.1",
-    "lucide-react": "^0.468.0",
     "mobx": "^5.15.4",
     "mobx-react": "^6",
     "mobx-state-tree": "^3.16.0",
@@ -91,11 +87,7 @@
     "react-window": "^1.8.9",
     "react-window-infinite-loader": "^1.0.5",
     "sanitize-html": "^2.12.1",
-    "shadcn": "^2.1.8",
-    "storybook": "^8.5.0",
     "strman": "^2.0.1",
-    "tailwind": "^4.0.0",
-    "tailwind-merge": "^2.6.0",
     "wavesurfer.js": "^6.0.1",
     "xpath-range": "^1.1.1",
     "zod": "^3.22.4"
@@ -115,14 +107,13 @@
     "@babel/preset-typescript": "^7.13.0",
     "@babel/runtime": "^7.12.13",
     "@biomejs/biome": "1.7.1",
-    "@chromatic-com/storybook": "^3",
     "@cypress/code-coverage": "^3.12.9",
     "@cypress/webpack-preprocessor": "^5.17.0",
     "@nx/cypress": "17.0.3",
     "@nx/jest": "17.0.3",
     "@nx/js": "17.2.8",
     "@nx/react": "17.0.3",
-    "@nx/storybook": "^20.3.2",
+    "@nx/storybook": "^17.2.8",
     "@nx/webpack": "17.0.3",
     "@nx/workspace": "17.0.3",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
@@ -130,11 +121,11 @@
     "@sentry/react": "^8.33",
     "@storybook/addon-essentials": "7.6.10",
     "@storybook/addon-interactions": "^7.5.3",
-    "@storybook/addon-styling-webpack": "1.0.0",
     "@storybook/core-server": "7.6.10",
+    "@storybook/jest": "^0.2.3",
     "@storybook/react-webpack5": "7.6.10",
-    "@storybook/test": "^8.5.0",
     "@storybook/test-runner": "^0.13.0",
+    "@storybook/testing-library": "^0.2.2",
     "@svgr/plugin-jsx": "^8.1.0",
     "@svgr/webpack": "^8.1.0",
     "@swc-node/register": "~1.6.7",
@@ -153,7 +144,6 @@
     "@types/sinon": "^17.0.3",
     "@types/strman": "^2.0.0",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.8.0",
-    "autoprefixer": "^10.4.20",
     "babel-jest": "^29.4.1",
     "babel-loader": "^8.2.2",
     "babel-plugin-import": "^1.13.0",
@@ -183,7 +173,7 @@
     "nx": "17.0.3",
     "pixelmatch": "^5.3.0",
     "pngjs": "^7.0.0",
-    "postcss": "^8.4.49",
+    "postcss": "^8.4.6",
     "postcss-loader": "^7.3.3",
     "proper-lockfile": "^4.1.2",
     "react-refresh": "^0.10.0",
@@ -196,8 +186,6 @@
     "style-loader": "^3.3.3",
     "stylelint": "^16.9.0",
     "stylelint-config-standard-scss": "^13.1.0",
-    "stylelint-config-tailwindcss": "^0.0.7",
-    "tailwindcss": "^3.4.16",
     "toml": "^3.0.0",
     "truncate-middle": "^1.0.6",
     "ts-jest": "^29.1.0",
