--- web/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./tsconfig.base.json"
-}
